requirements.txt

fastapi
uvicorn
numpy
onnxruntime
pillow
